# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trader — Tinkoff Invest API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Settings {
  hasToken: boolean;
  isSandbox: boolean;
  /** @nullable */
  accountId: string | null;
  maxOrderAmount: number;
  riskPercent: number;
  agentIntervalMinutes: number;
}

export interface UpdateSettingsBody {
  /** @nullable */
  token?: string | null;
  isSandbox?: boolean;
  /** @nullable */
  accountId?: string | null;
  /** @nullable */
  maxOrderAmount?: number | null;
  /** @nullable */
  riskPercent?: number | null;
  /** @nullable */
  agentIntervalMinutes?: number | null;
}

export interface Account {
  id: string;
  name: string;
  type: string;
  status: string;
}

export interface Position {
  figi: string;
  ticker: string;
  name: string;
  quantity: number;
  averagePrice: number;
  currentPrice: number;
  currentValue: number;
  unrealizedPnl: number;
  unrealizedPnlPercent: number;
  instrumentType: string;
}

export interface Portfolio {
  positions: Position[];
  totalAmountRub: number;
  expectedYield: number;
}

export interface PortfolioSummary {
  totalValue: number;
  totalPnl: number;
  totalPnlPercent: number;
  dailyPnl: number;
  dailyPnlPercent: number;
  positionsCount: number;
  cashRub: number;
}

export interface Instrument {
  figi: string;
  ticker: string;
  name: string;
  type: string;
  currency: string;
  exchange?: string;
  lot: number;
}

export interface Candle {
  time: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface LastPrice {
  figi: string;
  price: number;
  time: string;
}

export type OrderDirection =
  (typeof OrderDirection)[keyof typeof OrderDirection];

export const OrderDirection = {
  buy: "buy",
  sell: "sell",
} as const;

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  market: "market",
  limit: "limit",
} as const;

export interface Order {
  orderId: string;
  figi: string;
  ticker?: string;
  direction: OrderDirection;
  type: OrderType;
  quantity: number;
  /** @nullable */
  price?: number | null;
  status: string;
  createdAt: string;
}

export type PlaceOrderBodyDirection =
  (typeof PlaceOrderBodyDirection)[keyof typeof PlaceOrderBodyDirection];

export const PlaceOrderBodyDirection = {
  buy: "buy",
  sell: "sell",
} as const;

export type PlaceOrderBodyOrderType =
  (typeof PlaceOrderBodyOrderType)[keyof typeof PlaceOrderBodyOrderType];

export const PlaceOrderBodyOrderType = {
  market: "market",
  limit: "limit",
} as const;

export interface PlaceOrderBody {
  figi: string;
  direction: PlaceOrderBodyDirection;
  quantity: number;
  orderType: PlaceOrderBodyOrderType;
  /** @nullable */
  price?: number | null;
}

export type TradeLogEntryAction =
  (typeof TradeLogEntryAction)[keyof typeof TradeLogEntryAction];

export const TradeLogEntryAction = {
  buy: "buy",
  sell: "sell",
  hold: "hold",
  analyze: "analyze",
} as const;

export interface TradeLogEntry {
  id: number;
  figi: string;
  ticker: string;
  action: TradeLogEntryAction;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  price?: number | null;
  aiReasoning: string;
  success: boolean;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export interface TradeStats {
  totalTrades: number;
  successfulTrades: number;
  winRate: number;
  totalPnl: number;
  avgHoldingHours: number;
}

export interface AgentStatus {
  isRunning: boolean;
  /** @nullable */
  lastRunAt: string | null;
  /** @nullable */
  nextRunAt: string | null;
  /** @nullable */
  lastAction: string | null;
  totalAnalyses: number;
  totalTradesExecuted: number;
}

export interface AnalyzeBody {
  /** @nullable */
  figi?: string | null;
  executeIfConfident: boolean;
}

export interface WatchlistItem {
  id: number;
  figi: string;
  ticker: string;
  name: string;
  addedAt: string;
}

export interface AddWatchlistBody {
  figi: string;
  ticker: string;
  name: string;
}

export type SearchInstrumentsParams = {
  query?: string;
  type?: SearchInstrumentsType;
};

export type SearchInstrumentsType =
  (typeof SearchInstrumentsType)[keyof typeof SearchInstrumentsType];

export const SearchInstrumentsType = {
  stock: "stock",
  etf: "etf",
  bond: "bond",
  currency: "currency",
} as const;

export type GetCandlesParams = {
  interval?: GetCandlesInterval;
  from?: string;
  to?: string;
};

export type GetCandlesInterval =
  (typeof GetCandlesInterval)[keyof typeof GetCandlesInterval];

export const GetCandlesInterval = {
  "1min": "1min",
  "5min": "5min",
  "15min": "15min",
  hour: "hour",
  day: "day",
  week: "week",
  month: "month",
} as const;

export type ListTradeLogsParams = {
  limit?: number;
  offset?: number;
};