# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trader — Tinkoff Invest API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [portfolio]
      summary: Get portfolio positions from Tinkoff
      responses:
        "200":
          description: Portfolio positions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "400":
          description: API token not configured or error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /market/instruments:
    get:
      operationId: searchInstruments
      tags: [market]
      summary: Search instruments (stocks, ETFs)
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [stock, etf, bond, currency]
      responses:
        "200":
          description: List of instruments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instrument"
  /market/instruments/{figi}/candles:
    get:
      operationId: getCandles
      tags: [market]
      summary: Get price candles for an instrument
      parameters:
        - name: figi
          in: path
          required: true
          schema:
            type: string
        - name: interval
          in: query
          required: false
          schema:
            type: string
            enum: [1min, 5min, 15min, hour, day, week, month]
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: List of candles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
  /market/instruments/{figi}/price:
    get:
      operationId: getLastPrice
      tags: [market]
      summary: Get last price for an instrument
      parameters:
        - name: figi
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Last price
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LastPrice"
...
  /agent/analyze:
    post:
      operationId: analyzeAndTrade
      tags: [agent]
      summary: Trigger one AI analysis cycle (analyze market and optionally trade)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeBody"
      responses:
        "200":
          description: SSE stream of AI analysis and trade decision
          content:
            text/event-stream: {}
  /agent/watchlist:
    get:
      operationId: getWatchlist
      tags: [agent]
      summary: Get instruments the agent is watching
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addToWatchlist
      tags: [agent]
      summary: Add instrument to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWatchlistBody"
      responses:
        "201":
          description: Added to watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /agent/watchlist/{id}:
    delete:
      operationId: removeFromWatchlist
      tags: [agent]
      summary: Remove instrument from watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Portfolio:
      type: object
      properties:
        positions:
          type: array
          items:
            $ref: "#/components/schemas/Position"
        totalAmountRub:
          type: number
        expectedYield:
          type: number
      required:
        - positions
        - totalAmountRub
        - expectedYield
    Position:
      type: object
      properties:
        figi:
          type: string
        ticker:
          type: string
        name:
          type: string
        quantity:
          type: number
        averagePrice:
          type: number
        currentPrice:
          type: number
        currentValue:
          type: number
        unrealizedPnl:
          type: number
        unrealizedPnlPercent:
          type: number
        instrumentType:
          type: string
      required:
        - figi
        - ticker
        - name
        - quantity
        - averagePrice
        - currentPrice
        - currentValue
        - unrealizedPnl
        - unrealizedPnlPercent
        - instrumentType
...
    Instrument:
      type: object
      properties:
        figi:
          type: string
        ticker:
          type: string
        name:
          type: string
        type:
          type: string
        currency:
          type: string
        exchange:
          type: string
        lot:
          type: integer
      required:
        - figi
        - ticker
        - name
        - type
        - currency
        - lot
    Candle:
      type: object
      properties:
        time:
          type: string
          format: date-time
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: integer
      required:
        - time
        - open
        - high
        - low
        - close
        - volume
    LastPrice:
      type: object
      properties:
        figi:
          type: string
        price:
          type: number
        time:
          type: string
          format: date-time
      required:
        - figi
        - price
        - time
...
    AnalyzeBody:
      type: object
      properties:
        figi:
          type: ["string", "null"]
        executeIfConfident:
          type: boolean
      required:
        - executeIfConfident
    WatchlistItem:
      type: object
      properties:
        id:
          type: integer
        figi:
          type: string
        ticker:
          type: string
        name:
          type: string
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - figi
        - ticker
        - name
        - addedAt
    AddWatchlistBody:
      type: object
      properties:
        figi:
          type: string
        ticker:
          type: string
        name:
          type: string
      required:
        - figi
        - ticker
        - name